reminder/pyproject.toml
 [tool.poetry]
 name = "reminder-suite"
 version = "0.1.0"
 description = "MCP task scheduler server and Textual console agent"
 authors = ["Your Name <you@example.com>"]
 package-mode = false

 [tool.poetry.dependencies]
 python = "^3.11"
 fastapi = "^0.115.0"
 uvicorn = {extras = ["standard"], version = "^0.30.0"}
 pydantic = "^2.8.0"
sse-starlette = "^1.6.1"
 httpx = "^0.27.0"
 httpx-sse = "^0.4.0"
 python-dateutil = "^2.9.0.post0"
 textual = "^0.79.1"
 langchain-core = "^0.3.0"
 langchain = "^0.3.0"
 gigachain = "^0.1.6"
gigachat = "^0.1.0"
 python-dotenv = "^1.0.1"
 typing-extensions = "^4.12.2"

 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.0"

 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
